fixing 2 bugs : 1. Deleting: there seems to be a problem with deleting units, maybe in updating the indices. Example:
When you have a unit X, and inside of it a sub-unit Y, and inside of
it a sub-unit Z, deleting the unit Y causes selection to work
strangely: it doesn't work on Z, and when selecting Z inside the X
unit box, it selects the tokens in the Z unit box.
I think the model breaks, because when continuing the group units, it
sometimes creates empty unit boxes (see screenshot) and produces an
error.

2. Grouping units: to reproduce the bug -

a. create a unit from 4 tokens: t1,t2,t3,t4
b. create 4 sub-units, one from each token: u1,u2,u3,u4
c. group u1 and u2 together.